fix: enhance type safety and handle optional fields in database models and use cases
- Updated type annotations to use Optional for timestamps and model conversions.
- Improved handling of optional fields in various repository methods to prevent potential NoneType errors.
- Ensured consistent return types across repository methods for better type safety.

+++ src/infrastructure/database/models/load_model.py
@@ -140,7 +140,7 @@ class LoadModel(Base, TimestampMixin):
     def rate_per_mile(self) -> float:
         """Calculate rate per mile."""
         if self.miles and self.miles > 0:
-            return self.total_rate / self.miles
+            return float(self.total_rate / self.miles)
         return 0.0
 
     def __repr__(self) -> str:
